+++ packages/gov-ui-kit/src/components/avatars/avatarIcon/avatarIcon.test.tsx
@@ -1,0 +1,17 @@
+import { render, screen } from '@testing-library/react';
+
+import { IconType } from '../../icon';
+import { AvatarIcon, type IAvatarIconProps } from './avatarIcon';
+
+describe('<AvatarIcon /> component', () => {
+    const createTestComponent = (props?: Partial<IAvatarIconProps>) => {
+        const completeProps: IAvatarIconProps = { icon: IconType.ADD, ...props };
+
+        return <AvatarIcon {...completeProps} />;
+    };
+
+    it('renders the specified icon', () => {
+        render(createTestComponent({ icon: IconType.APP_FINANCE }));
+        expect(screen.getByTestId(IconType.APP_FINANCE)).toBeInTheDocument();
+    });
+});