+++ packages/gov-ui-kit/CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 -   Implement `InputDate` and `Avatar` components
+-   Add `AvatarIcon` documentation and tests
 
 ## [1.0.8] - 2024-01-17